Created a macro evalpoly2
This macro uses Horner's method to create an expression that
evaluates a polynomial. The difference between that and the
standard library macro @evalpoly is that in this case I don't
need to provide individual coefficients but the array (or tuple)
itself. The number of elements that should be used must be provided
as well. I think in the long run this will be better.
Since this is my first macro, I will actually use it later.