Started implementing the user interface to the library
For now I was using multiple dispatch on types that I previsously defined to specify humidity parameters (and other types of thermodynamic variables actually). Perhaps I could use an enum? Or single characters as I used to do. I will have to study a little bit to see what is the best interface.