The package is apparently working.
Including the wet bulb temperature related stuff. It agrees well with the C++/R version of the package.

I still have not written tests for these functions (user interface). This will probably be tricky But I have to do it somehow.

The code doesn't check temperature and pressure ranges. That is the next step.

The user interface doesn't have any examples or documentation.

Finally, there are several iterative procedures around the code and convergence should be checked and an exception thrown if the code fails to converge. Convergence should be testes for the entire range of the applicable pressures and temperatures.