Started working on units.
The Unitfil package is really nice and I'm trying to play along.

I could have implemented everything using units. In some respect this would be the best optio but it does have its drawbacks.

What if someone doesn't want to use Unitful? Maybe they are using it in a context where Unitful can not be used.

My solution is to use multiple dispatch to get it working. I am having some trouble so the interface is not working.