think the code using units is working.
There were some bugs leftover. It is easy to see that unit testing is important. There are so many places where things can go wrong...

Another major modification I made was to change the names of molar property. For a property that used to be called molarprop, now propm should be used. It saves a lot of typing.